chore(refactor): add internal/utils/ptrutil package with generic Deref/DerefOrZero/PtrNonZero; consolidate local ptr-helper functions (int32Ptr, int64Ptr, stringPtr, strPtr, boolPtr, ptrBool, intPtrToInt32, etc.) across dynamodb, neptunegraph, s3 store, wafv2, s3 services to use proto.* or ptrutil.*; migrate S3 store types (*int→*int32) for lifecycle/CORS/object-lock/replication fields; fix GetBucketReplication Priority omitempty (was always non-nil via &rule.Priority); gofmt sdk-tests cognito_srp.go

+++ internal/store/aws/s3/convert_bucket.go
@@ -295,7 +295,7 @@ func replicationRuleToProto(r *ReplicationRule) *pb.ReplicationRule {
 	}
 	return &pb.ReplicationRule{
 		Id:                      r.ID,
-		Priority:                int32(r.Priority),
+		Priority:                r.Priority,
 		Status:                  r.Status,
 		Filter:                  replicationFilterToProto(r.Filter),
 		Destination:             replicationDestinationToProto(r.Destination),
@@ -309,7 +309,7 @@ func protoToReplicationRule(p *pb.ReplicationRule) *ReplicationRule {
 	}
 	return &ReplicationRule{
 		ID:                      p.Id,
-		Priority:                int(p.Priority),
+		Priority:                p.Priority,
 		Status:                  p.Status,
 		Filter:                  protoToReplicationFilter(p.Filter),
 		Destination:             protoToReplicationDestination(p.Destination),
